Refactor backend and extract globally used methods to common packages
