+++ src/app/backend/resource/event/event.go
@@ -18,6 +18,7 @@ import (
 	"strings"
 
 	"github.com/kubernetes/dashboard/resource/common"
+
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/types"
 )
