+++ src/app/backend/resource/event/eventcommon.go
@@ -16,6 +16,7 @@ package event
 
 import (
 	"github.com/kubernetes/dashboard/resource/common"
+
 	"k8s.io/kubernetes/pkg/api"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/fields"
