+++ src/app/backend/resource/namespace/namespace.go
@@ -17,7 +17,7 @@ package namespace
 import (
 	"log"
 
-	api "k8s.io/kubernetes/pkg/api"
+	"k8s.io/kubernetes/pkg/api"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/fields"
 	"k8s.io/kubernetes/pkg/labels"
