+++ src/app/backend/resource/pod/podlist.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/kubernetes/dashboard/client"
 	"github.com/kubernetes/dashboard/resource/common"
+
 	"k8s.io/kubernetes/pkg/api"
 	k8sClient "k8s.io/kubernetes/pkg/client/unversioned"
 )
