+++ src/app/backend/resource/replicationcontroller/replicationcontrollercommon.go
@@ -16,6 +16,7 @@ package replicationcontroller
 
 import (
 	"github.com/kubernetes/dashboard/resource/common"
+
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
@@ -114,13 +115,4 @@ func getServicesForDeletion(client client.Interface, labelSelector labels.Select
 	}
 
 	return services.Items, nil
-}
-
-// GetContainerImages returns container image strings from the given pod spec.
-func GetContainerImages(podTemplate *api.PodSpec) []string {
-	var containerImages []string
-	for _, container := range podTemplate.Containers {
-		containerImages = append(containerImages, container.Image)
-	}
-	return containerImages
 }
