+++ src/app/backend/resource/replicationcontroller/replicationcontrollerevents.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/kubernetes/dashboard/resource/common"
 	resourceEvent "github.com/kubernetes/dashboard/resource/event"
+
 	"k8s.io/kubernetes/pkg/api"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 )
