+++ src/test/backend/resource/common/endpoint_test.go
@@ -100,6 +100,31 @@ func TestGetExternalEndpoints(t *testing.T) {
 	}
 }
 
+func TestGetInternalEndpoint(t *testing.T) {
+	cases := []struct {
+		serviceName, namespace string
+		ports                  []api.ServicePort
+		expected               Endpoint
+	}{
+		{"my-service", api.NamespaceDefault, nil, Endpoint{Host: "my-service"}},
+		{"my-service", api.NamespaceDefault,
+			[]api.ServicePort{{Name: "foo", Port: 8080, Protocol: "TCP"}},
+			Endpoint{Host: "my-service", Ports: []ServicePort{{Port: 8080, Protocol: "TCP"}}}},
+		{"my-service", "my-namespace", nil, Endpoint{Host: "my-service.my-namespace"}},
+		{"my-service", "my-namespace",
+			[]api.ServicePort{{Name: "foo", Port: 8080, Protocol: "TCP"}},
+			Endpoint{Host: "my-service.my-namespace",
+				Ports: []ServicePort{{Port: 8080, Protocol: "TCP"}}}},
+	}
+	for _, c := range cases {
+		actual := GetInternalEndpoint(c.serviceName, c.namespace, c.ports)
+		if !reflect.DeepEqual(actual, c.expected) {
+			t.Errorf("getInternalEndpoint(%+v, %+v, %+v) == %+v, expected %+v",
+				c.serviceName, c.namespace, c.ports, actual, c.expected)
+		}
+	}
+}
+
 func TestGetNodePortEndpoints(t *testing.T) {
 	cases := []struct {
 		pods     []api.Pod
